requirement.txt
django-shopping-cart